Fix test.
This test was cherry-picked from 1.2 where GradleTestProject
behaves slightly differently. In particular when creating from
a testProject, the base build.gradle with the buildSrc section
is not created in 1.1 (but is in 1.2).

This made the test fail since there was no buildSrc section.

This fixes the issue, but require a slight tweak to the API
of GradleTestProject which is not worth putting in 1.2 at
the moment.

Change-Id: If2230de51a664fada46a6dbb8a8b20b08b1d2c6a
(cherry picked from commit 7621d82221204be5ecb8ebd588d9f71a8bac3e83)